JIRA JAH-17 Modify maven build for mon-thresh that creates a jar that can be used for a production build.
Needed to strip out the storm jar from the consolidated jar.

Needed to remove storm.yaml because Storm complains. Moved the registration of the Serializer to ThresholdingEngine.

Added the storm-core jar to the deb so it can be used for the local mode in mini-mon

Added logback.xml to the deb since storm-core.jar also has a logback.xml and that confuses logback. Ensure our logback.xml is used.

Added the storm-core.jar and logback.xml to the start of thresh in the deb

Had to rework how the AlarmEventForwarder was injected into the AlarmThresholdingBolt because the old way didn't work in a Storm cluster because the TopologyModule wasn't loaded on the worker when prepare was called.