Use external Zookeeper in Local mode
By default Storm uses the in-process ZooKeeper when the local mode
is chosen.

The in-process ZooKeeper doesn't handle the trim of ZooKeeper
Transactional logs & snapshots.

There is a constructor which enables the possibility to use an
external ZooKeeper when the local mode is used. [1][2]

[1] https://issues.apache.org/jira/browse/STORM-213
[2] https://github.com/apache/storm/pull/137/files

Change-Id: I4c6e6f1dbacf093f94b566abc7251f4e33eff93f
(cherry picked from commit d156980a528caf1797df1a621719dd76668a18dd)